fix: Tab 오동작 수정 및 애니메이션 추가
