+++ Sources/Montage/Common/InteractionDisablingViewModifier.swift
@@ -1,0 +1,18 @@
+//
+//  InteractionDisablingViewModifier.swift
+//  Montage
+//
+//  Created by 김삼열 on 11/21/24.
+//
+
+import SwiftUI
+
+struct InteractionDisablingViewModifier: ViewModifier {
+    let disabled: Bool
+    
+    func body(content: Content) -> some View {
+        content
+            .allowsHitTesting(!disabled)
+            .disableSwipeBack(disabled)
+    }
+}
